Update main.py
First functional version - Works as a basic timer with microsecond precision. However, this precision is unreal as the error of the clock is higher than microsecond. That's because we are updating it only every ms. The definitive version will only show the numbers until the cs to be less distracting while playing and to show only accuracy timing.